perf tests hwmon_pmu: Use PRIu64 + (uint64_t) cast for a __u64 field to work more widely
While testing perf with an updated Debian experimental cross compiler
(gcc version 14.2.0 (Debian 14.2.0-13)) this started failing:

  In file included from tests/hwmon_pmu.c:12:
  tests/hwmon_pmu.c: In function 'do_test':
  tests/hwmon_pmu.c:199:34: error: format '%lld' expects argument of type 'long long int', but argument 7 has type '__u64' {aka 'long unsigned int'} [-Werror=format=]
    199 |                         pr_debug("FAILED %s:%d Unexpected config for '%s', %lld != %ld\n",
        |                                  ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  /git/perf-7.1.0-rc5/tools/perf/util/debug.h:20:21: note: in definition of macro 'pr_fmt'
     20 | #define pr_fmt(fmt) fmt
        |                     ^~~
  tests/hwmon_pmu.c:199:25: note: in expansion of macro 'pr_debug'
    199 |                         pr_debug("FAILED %s:%d Unexpected config for '%s', %lld != %ld\n",
        |                         ^~~~~~~~
  tests/hwmon_pmu.c:199:79: note: format string is defined here
    199 |                         pr_debug("FAILED %s:%d Unexpected config for '%s', %lld != %ld\n",
        |                                                                            ~~~^
        |                                                                               |
        |                                                                               long long int
        |                                                                            %ld
    LD      /tmp/build/perf/util/intel-pt-decoder/perf-util-in.o

The usual make that %lld a PRIu64 (since arg7 is
evsel->core.attr.config, which is a __u64) but then on Fedora 44 (gcc
version 16.1.1 20260515 (Red Hat 16.1.1-2)) it ends up with:

  In file included from tests/hwmon_pmu.c:13:
  tests/hwmon_pmu.c: In function ‘do_test’:
  tests/hwmon_pmu.c:200:34: error: format ‘%lu’ expects argument of type ‘long unsigned int’, but argument 7 has type ‘__u64’ {aka ‘long long unsigned int’} [-Werror=format=]
    200 |                         pr_debug("FAILED %s:%d Unexpected config for '%s', %" PRIu64 " != %ld\n",
        |                                  ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  /home/acme/git/perf-tools-next2/tools/perf/util/debug.h:20:21: note: in definition of macro ‘pr_fmt’
     20 | #define pr_fmt(fmt) fmt
        |                     ^~~
  tests/hwmon_pmu.c:200:25: note: in expansion of macro ‘pr_debug’
    200 |                         pr_debug("FAILED %s:%d Unexpected config for '%s', %" PRIu64 " != %ld\n",
        |                         ^~~~~~~~
  cc1: all warnings being treated as errors

So the way to satisfy both compilers is to also add a (u64) cast to
arg7.

Cc: Adrian Hunter <adrian.hunter@intel.com>
Cc: Ian Rogers <irogers@google.com>
Cc: James Clark <james.clark@linaro.org>
Cc: Jiri Olsa <jolsa@kernel.org>
Cc: Namhyung Kim <namhyung@kernel.org>

+++ tools/perf/tests/hwmon_pmu.c
@@ -2,6 +2,7 @@
 #include "hwmon_pmu.h"
 
 #include <errno.h>
+#include <inttypes.h>
 
 #include <fcntl.h>
 #include <linux/compiler.h>
@@ -196,9 +197,9 @@ static int do_test(size_t i, bool with_pmu, bool with_alias)
 			continue;
 
 		if (evsel->core.attr.config != (u64)test_events[i].key.type_and_num) {
-			pr_debug("FAILED %s:%d Unexpected config for '%s', %lld != %ld\n",
+			pr_debug("FAILED %s:%d Unexpected config for '%s', %" PRIu64 " != %ld\n",
 				__FILE__, __LINE__, str,
-				evsel->core.attr.config,
+				(uint64_t)evsel->core.attr.config,
 				test_events[i].key.type_and_num);
 			ret = TEST_FAIL;
 			goto out;
